+++ tests/test_mpisystem.cpp
@@ -57,7 +57,7 @@ BOOST_AUTO_TEST_CASE(test_1, *boost::unit_test::timeout(60)) {
       // check for linear scaling (grace 20%)
       if (TestHelper::getRank(MPI_COMM_WORLD) == 0) {
         BOOST_TEST(static_cast<double>(vmSizes[i]) <=
-                   (vmSizes[0] * groupSizes[i] / groupSizes[0] * 1.2));
+                   (static_cast<double>(vmSizes[0]) * groupSizes[i] / groupSizes[0] * 1.2));
       }
     }
   }
